requirements.txt
asgiref==3.2.7
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-select2==7.2.3
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
